backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
requests==2.32.3
sentence-transformers==2.7.0
google-generativeai==0.5.4
pdfplumber==0.10.4
numpy==1.26.4
pymongo==4.6.3        
dnspython==2.4.2      